test: extract common part out of test scenarios
